LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/size-feature-evaluation-expected.txt

Test

PASS (width < 100px) (.horizontal)
FAIL (width >= 100px) (.horizontal) assert_equals: expected "true" but got ""
FAIL (min-width: 100px) (.horizontal) assert_equals: expected "true" but got ""
PASS (min-width: 101px) (.horizontal)
FAIL (max-width: 100px) (.horizontal) assert_equals: expected "true" but got ""
PASS (max-width: 99px) (.horizontal)
PASS (height < 200px) (.horizontal)
FAIL (height >= 200px) (.horizontal) assert_equals: expected "true" but got ""
FAIL (min-height: 200px) (.horizontal) assert_equals: expected "true" but got ""
PASS (min-height: 201px) (.horizontal)
FAIL (max-height: 200px) (.horizontal) assert_equals: expected "true" but got ""
PASS (max-height: 199px) (.horizontal)
PASS (inline-size < 100px) (.horizontal)
FAIL (inline-size >= 100px) (.horizontal) assert_equals: expected "true" but got ""
FAIL (min-inline-size: 100px) (.horizontal) assert_equals: expected "true" but got ""
PASS (min-inline-size: 101px) (.horizontal)
FAIL (max-inline-size: 100px) (.horizontal) assert_equals: expected "true" but got ""
PASS (max-inline-size: 99px) (.horizontal)
PASS (block-size < 200px) (.horizontal)
FAIL (block-size >= 200px) (.horizontal) assert_equals: expected "true" but got ""
FAIL (min-block-size: 200px) (.horizontal) assert_equals: expected "true" but got ""
PASS (min-block-size: 201px) (.horizontal)
FAIL (max-block-size: 200px) (.horizontal) assert_equals: expected "true" but got ""
PASS (max-block-size: 199px) (.horizontal)
PASS (orientation: landscape) (.horizontal)
FAIL (orientation: portrait) (.horizontal) assert_equals: expected "true" but got ""
FAIL (aspect-ratio: 1/2) (.horizontal) assert_equals: expected "true" but got ""
PASS (aspect-ratio: 2/1) (.horizontal)
PASS (width < 100px) (.vertical)
FAIL (width >= 100px) (.vertical) assert_equals: expected "true" but got ""
FAIL (min-width: 100px) (.vertical) assert_equals: expected "true" but got ""
PASS (min-width: 101px) (.vertical)
FAIL (max-width: 100px) (.vertical) assert_equals: expected "true" but got ""
PASS (max-width: 99px) (.vertical)
PASS (height < 200px) (.vertical)
FAIL (height >= 200px) (.vertical) assert_equals: expected "true" but got ""
FAIL (min-height: 200px) (.vertical) assert_equals: expected "true" but got ""
PASS (min-height: 201px) (.vertical)
FAIL (max-height: 200px) (.vertical) assert_equals: expected "true" but got ""
PASS (max-height: 199px) (.vertical)
PASS (block-size < 100px) (.vertical)
FAIL (block-size >= 100px) (.vertical) assert_equals: expected "true" but got ""
FAIL (min-block-size: 100px) (.vertical) assert_equals: expected "true" but got ""
PASS (min-block-size: 101px) (.vertical)
FAIL (max-block-size: 100px) (.vertical) assert_equals: expected "true" but got ""
PASS (max-block-size: 99px) (.vertical)
PASS (inline-size < 200px) (.vertical)
FAIL (inline-size >= 200px) (.vertical) assert_equals: expected "true" but got ""
FAIL (min-inline-size: 200px) (.vertical) assert_equals: expected "true" but got ""
PASS (min-inline-size: 201px) (.vertical)
FAIL (max-inline-size: 200px) (.vertical) assert_equals: expected "true" but got ""
PASS (max-inline-size: 199px) (.vertical)
PASS (orientation: landscape) (.vertical)
FAIL (orientation: portrait) (.vertical) assert_equals: expected "true" but got ""
FAIL (aspect-ratio: 1/2) (.vertical) assert_equals: expected "true" but got ""
PASS (aspect-ratio: 2/1) (.vertical)
